Require parents to have been previously seen to skip new element recolor.
An issue with the previous implementation could be reproduced if a dark
background element (with no color definition and no text) has a light
background child added dynamically (caniuse.com has this at this time). The
previous algorithm would find the parent with a defined color and assume it
has already been fixed, even though it has not (due to passing the contrast
check). New algorithm requires existence of custom data attribute to ensure
contrast is good.

Also added a test for this case in a separate test file for non-input
elements.