Make sure handling of coloring cases are exclusive of each other.
Consider a page that has a defined background image and color, but the
color is dark, so there is no contrast issue. The original code would skip
out of the case for setting a foreground (since contrast is ok), but would
then enter the background image case and override the background color
unnecessarily. By making the case checking an if-else ladder, this issue
can be avoided.