Fix background script-tab messaging:
The previous connection based method used a single port with the background
script, which meant only the last connected tab could talk to it, breaking
the toggle operation in all other tabs.

Fix uses the simple message passing method, and gets the tab id of the
active tab to specify the target of the toggle message.