Undo removal of contrast check for input elements.
Sites that naturally use a dark theme can set input [bg]colors to be
coincidentally the same as the user's dark theme (e.g. common with
Adwaita-dark which uses a foreground color of white). The result is that
the extension mistakenly thinks that the property has not been set. To
avoid issues, the extension must do the contrast check on all elements,
as there is currently no way for the extension to tell that an author has
explicitly set a property that happens to match the theme style.