Delay messaging to embeds, to ensure receiving script has loaded.
GH-23 occurs because the signal for the iframe to fix its elements is sent
before the embed.js file has run and set up the receiving listener. This is
exasperated by 50b207da, requiring that the scripts load when the DOM is
finished loading. To compensate, added a delay to sending the message to
minimize the chances of the out-of-order execution. This is still a hacky
fix; a more robust mechanism of ensuring message receipt will need to be
added later.